Separate clearAll function to another file to fix linters
