+++ src/module/delete.js
@@ -1,5 +1,4 @@
 import { indexUpdate } from './crud.js';
-import display from './display.js';
 import getFromStorage, { saveInStorage } from './storage.js';
 
 const removeItem = (list, task, display) => {
@@ -9,12 +8,4 @@ const removeItem = (list, task, display) => {
   display(getFromStorage());
 };
 
-const clearAll = (list) => {
-  list = list.filter((task) => task.completed !== true);
-  saveInStorage(list);
-  list = indexUpdate(getFromStorage());
-  display(list);
-}
-
 export default removeItem;
-export { clearAll };
